+++ CHANGES.rst
@@ -5,7 +5,8 @@ Changelog
 2.2.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Better multilingual profile
+  [sneridagh]
 
 
 2.2.0 (2021-04-21)